pyproject.toml
[project]
name = "hirnu"
version = "0.1.0"
description = "Model training pipeline for Hirnu language using Apple MLX Framework"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mlx-lm>=0.28.3",
    "pandas>=2.2.0",
    "huggingface-hub>=0.26.0",
    "pyyaml>=6.0",
    "tqdm>=4.67.0",
    "numpy>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.ruff]
line-length = 100
target-version = "py313"
